+++ test/contstants.test.js
@@ -47,6 +47,10 @@ describe('constants', function() {
     assert.ok(isNaN(math.NaN));
   });
 
+  it('should have null', function() {
+    assert.strictEqual(math['null'], null);
+  });
+
   it('should have version number', function() {
     assert.equal(math.version, require('../package.json').version);
   });